\name{print.epGPCA.inference.battery}
\alias{print.epGPCA.inference.battery}
\title{Print Generalized Principal Components Analysis (GPCA) Inference results}
\description{
	Print Generalized Principal Components Analysis (GPCA) Inference results.
}
\usage{
	\method{print}{epGPCA.inference.battery}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the epGPCA.inference.battery class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
