% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inpoBoot.R
\name{print.inpoBoot}
\alias{print.inpoBoot}
\title{Print results from InPosition Bootstraps}
\usage{
\method{print}{inpoBoot}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the inpoBoot class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print bootstrap results from the InPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
