% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inpoBootTests.R
\name{print.inpoBootTests}
\alias{print.inpoBootTests}
\title{Print results from InPosition Bootstrap Ratio Tests}
\usage{
\method{print}{inpoBootTests}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the inpoBootTests class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print bootstrap ratio tests results from the InPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
