\name{print.inpoComponents}
\alias{print.inpoComponents}
\title{Print results from InPosition Components Permutation Test}
\description{
	Print Components permutation test results from the inposition.
}
\usage{
	\method{print}{inpoComponents}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the inpoComponents class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
