\name{print.inpoOmni}
\alias{print.inpoOmni}
\title{Print results from InPosition Omnibus Permutation Test}
\description{
	Print Omnibus permutation test results from the inposition.
}
\usage{
	\method{print}{inpoOmni}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the inpoOmni class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
