\name{print.inpoOutput}
\alias{print.inpoOutput}
\title{Print results from InPosition}
\description{
	Print results from the InPosition.
}
\usage{
	\method{print}{inpoOutput}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the inpoOutput class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\seealso{
	\code{\link{epPCA.inference.battery}}, \code{\link{inGraphs}}
}
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
