\name{print.epCA.inference.battery}
\alias{print.epCA.inference.battery}
\title{Print Correspondence Analysis (CA) Inference results}
\description{
	Print Correspondence Analysis (CA) Inference results.
}
\usage{
	\method{print}{epCA.inference.battery}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the epCA.inference.battery class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
