% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epMCA.inference.battery.R
\name{print.epMCA.inference.battery}
\alias{print.epMCA.inference.battery}
\title{Print Multiple Correspondence Analysis (MCA) Inference results}
\usage{
\method{print}{epMCA.inference.battery}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the
epMCA.inference.battery class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Multiple Correspondence Analysis (MCA) Inference results.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
