% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epPCA.inference.battery.R
\name{print.epPCA.inference.battery}
\alias{print.epPCA.inference.battery}
\title{Print Principal Components Analysis (PCA) Inference results}
\usage{
\method{print}{epPCA.inference.battery}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the
epPCA.inference.battery class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Principal Components Analysis (PCA) Inference results.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
