\name{print.inpoBoot}
\alias{print.inpoBoot}
\title{Print results from InPosition Bootstraps}
\description{
	Print bootstrap results from the InPosition.
}
\usage{
	\method{print}{inpoBoot}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the inpoBoot class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
