\name{print.inpoBootTests}
\alias{print.inpoBootTests}
\title{Print results from InPosition Bootstrap Ratio Tests}
\description{
	Print bootstrap ratio tests results from the InPosition.
}
\usage{
	\method{print}{inpoBootTests}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the inpoBootTests class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
