% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inpoComponents.R
\name{print.inpoComponents}
\alias{print.inpoComponents}
\title{Print results from InPosition Components Permutation Test}
\usage{
\method{print}{inpoComponents}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the inpoComponents class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Components permutation test results from the inposition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
