% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inpoOmni.R
\name{print.inpoOmni}
\alias{print.inpoOmni}
\title{Print results from InPosition Omnibus Permutation Test}
\usage{
\method{print}{inpoOmni}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the inpoOmni class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Omnibus permutation test results from the inposition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
