% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inpoOutput.R
\name{print.inpoOutput}
\alias{print.inpoOutput}
\title{Print results from InPosition}
\usage{
\method{print}{inpoOutput}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the inpoOutput class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print results from the InPosition.
}
\seealso{
\code{\link{epPCA.inference.battery}}, \code{\link{inGraphs}}
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
